# api/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Mutual Fund Data API
  description: High-performance, locally cached REST API retrieving and indexing Indian Mutual Fund scheme metadata and historical Net Asset Value (NAV) records from mfapi.in.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
  /api/v1/schemes:
    get:
      summary: Query and search scheme catalog
      description: Paginated listing and filtering of mutual fund schemes by name search, fund house, scheme type, and category.
      operationId: getSchemes
      parameters:
        - name: page
          in: query
          required: false
          description: Zero-based page index (default 0)
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          description: Number of records per page (default 50, maximum 200)
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - name: search
          in: query
          required: false
          description: Case-insensitive search keyword matched against scheme name
          schema:
            type: string
        - name: fundHouse
          in: query
          required: false
          description: Exact or partial filter by fund house name
          schema:
            type: string
        - name: schemeType
          in: query
          required: false
          description: Filter by scheme type (e.g., 'Open Ended Schemes')
          schema:
            type: string
        - name: schemeCategory
          in: query
          required: false
          description: Filter by scheme category (e.g., 'Equity Scheme - Small Cap Fund')
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved paginated schemes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemePageResponse'
              examples:
                sampleCatalogResponse:
                  value:
                    content:
                      - schemeCode: 125497
                        schemeName: "SBI SMALL CAP FUND - Direct Plan - Growth"
                        fundHouse: "SBI Mutual Fund"
                        schemeType: "Open Ended Schemes"
                        schemeCategory: "Equity Scheme - Small Cap Fund"
                        isinGrowth: "INF200K01T51"
                        isinDivReinvestment: null
                        navSynced: true
                    page:
                      number: 0
                      size: 50
                      totalElements: 37860
                      totalPages: 758
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/schemes/{scheme_code}/nav:
    get:
      summary: Retrieve historical NAV records for a scheme
      description: Returns scheme metadata and date-ordered historical NAV points. Lazy-loads and permanently caches NAV data from mfapi.in on first request if not already synced.
      operationId: getSchemeNav
      parameters:
        - name: scheme_code
          in: path
          required: true
          description: Unique integer code identifying the mutual fund scheme
          schema:
            type: integer
        - name: startDate
          in: query
          required: false
          description: Filter NAV records on or after this date (ISO-8601 format YYYY-MM-DD)
          schema:
            type: string
            format: date
            pattern: '^\d{4}-\d{2}-\d{2}$'
        - name: endDate
          in: query
          required: false
          description: Filter NAV records on or before this date (ISO-8601 format YYYY-MM-DD)
          schema:
            type: string
            format: date
            pattern: '^\d{4}-\d{2}-\d{2}$'
        - name: sort
          in: query
          required: false
          description: Sort ordering of NAV records by date
          schema:
            type: string
            enum:
              - desc
              - asc
            default: desc
        - name: forceRefresh
          in: query
          required: false
          description: Force re-fetching NAV data from upstream regardless of cached state
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Historical NAV data and scheme metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NavHistoryResponse'
              examples:
                sampleNavResponse:
                  value:
                    meta:
                      schemeCode: 125497
                      schemeName: "SBI SMALL CAP FUND - Direct Plan - Growth"
                      fundHouse: "SBI Mutual Fund"
                      schemeType: "Open Ended Schemes"
                      schemeCategory: "Equity Scheme - Small Cap Fund"
                      isinGrowth: "INF200K01T51"
                      isinDivReinvestment: null
                      totalPoints: 2
                      startDate: "2026-09-03"
                      endDate: "2026-09-04"
                    data:
                      - date: "2026-09-04"
                        nav: 216.6525
                      - date: "2026-09-03"
                        nav: 216.4037
        '400':
          description: Bad request (e.g. invalid date format or startDate is after endDate)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                badRequest:
                  value:
                    code: "INVALID_DATE_RANGE"
                    message: "startDate (2026-09-10) cannot be after endDate (2026-09-01)"
                    timestamp: "2026-09-06T11:42:00Z"
        '404':
          description: Scheme not found locally or upstream
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  value:
                    code: "SCHEME_NOT_FOUND"
                    message: "Mutual fund scheme with code 999999 was not found"
                    timestamp: "2026-09-06T11:42:00Z"
        '502':
          description: Bad Gateway (upstream mfapi.in request failed or returned unexpected response)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                badGateway:
                  value:
                    code: "UPSTREAM_GATEWAY_ERROR"
                    message: "Failed to retrieve NAV data from upstream provider mfapi.in"
                    timestamp: "2026-09-06T11:42:00Z"

  /api/v1/admin/schemes/sync:
    post:
      summary: Synchronize mutual fund scheme catalog
      description: Ingests the complete master directory of mutual fund schemes from mfapi.in/mf/latest and bulk upserts them into the local database.
      operationId: syncSchemes
      responses:
        '200':
          description: Catalog synchronization completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncSummaryResponse'
              examples:
                sampleSyncSummary:
                  value:
                    status: "COMPLETED"
                    totalFetched: 37860
                    inserted: 37860
                    updated: 0
                    durationMs: 1800
                    syncedAt: "2026-09-06T11:15:00Z"
        '502':
          description: Bad Gateway (upstream catalog fetch failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                badGateway:
                  value:
                    code: "UPSTREAM_GATEWAY_ERROR"
                    message: "Failed to download master scheme directory from mfapi.in"
                    timestamp: "2026-09-06T11:42:00Z"

components:
  schemas:
    SchemeResponse:
      type: object
      description: Mutual fund scheme metadata item
      required:
        - schemeCode
        - schemeName
        - navSynced
      properties:
        schemeCode:
          type: integer
          description: Unique scheme identification code
          example: 125497
        schemeName:
          type: string
          description: Full mutual fund scheme name
          example: "SBI SMALL CAP FUND - Direct Plan - Growth"
        fundHouse:
          type:
            - string
            - "null"
          description: Asset Management Company / Mutual Fund House
          example: "SBI Mutual Fund"
        schemeType:
          type:
            - string
            - "null"
          description: Scheme type classification
          example: "Open Ended Schemes"
        schemeCategory:
          type:
            - string
            - "null"
          description: SEBI category classification
          example: "Equity Scheme - Small Cap Fund"
        isinGrowth:
          type:
            - string
            - "null"
          description: ISIN code for growth option
          example: "INF200K01T51"
        isinDivReinvestment:
          type:
            - string
            - "null"
          description: ISIN code for dividend reinvestment option
          example: null
        navSynced:
          type: boolean
          description: Whether NAV records for this scheme have been synced locally
          example: true

    PageMetadata:
      type: object
      description: Pagination metadata details
      required:
        - number
        - size
        - totalElements
        - totalPages
      properties:
        number:
          type: integer
          description: Current page index (0-based)
          example: 0
        size:
          type: integer
          description: Page size
          example: 50
        totalElements:
          type: integer
          format: int64
          description: Total number of matching elements across all pages
          example: 37860
        totalPages:
          type: integer
          description: Total number of pages
          example: 758

    SchemePageResponse:
      type: object
      description: Paginated response envelope for scheme listings
      required:
        - content
        - page
      properties:
        content:
          type: array
          description: List of schemes on the current page
          items:
            $ref: '#/components/schemas/SchemeResponse'
        page:
          $ref: '#/components/schemas/PageMetadata'

    NavPoint:
      type: object
      description: Daily Net Asset Value (NAV) record
      required:
        - date
        - nav
      properties:
        date:
          type: string
          format: date
          pattern: '^\d{4}-\d{2}-\d{2}$'
          description: Date of Net Asset Value formatted as YYYY-MM-DD
          example: "2026-09-04"
        nav:
          type: number
          format: float
          description: Net Asset Value per unit in INR
          example: 216.6525

    NavMeta:
      type: object
      description: Metadata header for historical NAV query response
      required:
        - schemeCode
        - schemeName
        - totalPoints
      properties:
        schemeCode:
          type: integer
          description: Unique scheme identification code
          example: 125497
        schemeName:
          type: string
          description: Scheme name
          example: "SBI SMALL CAP FUND - Direct Plan - Growth"
        fundHouse:
          type:
            - string
            - "null"
          description: Fund house name
          example: "SBI Mutual Fund"
        schemeType:
          type:
            - string
            - "null"
          description: Scheme type
          example: "Open Ended Schemes"
        schemeCategory:
          type:
            - string
            - "null"
          description: Scheme category
          example: "Equity Scheme - Small Cap Fund"
        isinGrowth:
          type:
            - string
            - "null"
          description: Growth option ISIN
          example: "INF200K01T51"
        isinDivReinvestment:
          type:
            - string
            - "null"
          description: Dividend reinvestment ISIN
          example: null
        totalPoints:
          type: integer
          description: Count of NAV data points returned in the response
          example: 2
        startDate:
          type:
            - string
            - "null"
          format: date
          description: Earliest date included in response range (YYYY-MM-DD)
          example: "2026-09-03"
        endDate:
          type:
            - string
            - "null"
          format: date
          description: Latest date included in response range (YYYY-MM-DD)
          example: "2026-09-04"

    NavHistoryResponse:
      type: object
      description: Response containing scheme metadata and historical NAV timeseries
      required:
        - meta
        - data
      properties:
        meta:
          $ref: '#/components/schemas/NavMeta'
        data:
          type: array
          description: Chronologically ordered NAV data points
          items:
            $ref: '#/components/schemas/NavPoint'

    SyncSummaryResponse:
      type: object
      description: Metrics and summary report of admin scheme catalog ingestion
      required:
        - status
        - totalFetched
        - inserted
        - updated
        - durationMs
        - syncedAt
      properties:
        status:
          type: string
          description: Status of synchronization operation (e.g. COMPLETED)
          example: "COMPLETED"
        totalFetched:
          type: integer
          description: Total schemes fetched from upstream
          example: 37860
        inserted:
          type: integer
          description: Total new scheme records inserted into database
          example: 37860
        updated:
          type: integer
          description: Total existing scheme records updated in database
          example: 0
        durationMs:
          type: integer
          format: int64
          description: Ingestion runtime duration in milliseconds
          example: 1800
        syncedAt:
          type: string
          format: date-time
          description: ISO-8601 UTC timestamp of ingestion completion
          example: "2026-09-06T11:15:00Z"

    ErrorResponse:
      type: object
      description: Standard unified error response payload
      required:
        - code
        - message
        - timestamp
      properties:
        code:
          type: string
          description: Error classification code
          example: "SCHEME_NOT_FOUND"
        message:
          type: string
          description: Human-readable error description
          example: "Mutual fund scheme with code 999999 was not found"
        timestamp:
          type: string
          format: date-time
          description: ISO-8601 timestamp when error occurred
          example: "2026-09-06T11:42:00Z"